refactor: utils 2020 file
